refactor (internal/taskmanager): split the taskmanager into 2 services: dispatcher and monitor
dispatcher creates the data in redis. and assigns performers

monitor updates the data in redis and db upon completion/failure
